(READY TO MERGE) Hibernate Upgrade Code

Approved-by: Dennis Warren

+++ src/main/java/org/oscarehr/provider/web/ProviderPropertyAction.java
@@ -433,7 +433,7 @@ public class ProviderPropertyAction extends DispatchAction {
         }
         if(mode.equals("new")){
             //save and get most recent id
-            QueueDao queueDao = (QueueDao) SpringUtils.getBean("queueDao");
+            QueueDao queueDao = (QueueDao) SpringUtils.getBean(QueueDao.class);
             queueDao.addNewQueue(defaultQ);
             String lastId=queueDao.getLastId();
             prop.setValue(lastId);
@@ -465,7 +465,7 @@ public class ProviderPropertyAction extends DispatchAction {
         if(prop==null){
             prop=new UserProperty();
         }
-        QueueDao queueDao = (QueueDao) SpringUtils.getBean("queueDao");
+        QueueDao queueDao = (QueueDao) SpringUtils.getBean(QueueDao.class);
         List<Hashtable> queues= queueDao.getQueues();
         Collection<LabelValueBean> viewChoices=new ArrayList<LabelValueBean>();
         viewChoices.add(new LabelValueBean("None","-1"));
@@ -2066,12 +2066,22 @@ public ActionForward viewEDocBrowserInDocumentReport(ActionMapping actionmapping
 		property.setName(UserProperty.TICKLER_TASK_ASSIGNEE);
 	}
 
-	if(delete){
-	 userPropertyDAO.delete(property);
-	}else{
-	 property.setValue(tickerTaskAssignee);
-	 userPropertyDAO.saveProp(property);
-	}
+	try {
+        if (delete) {
+            if (property.getId() != null) {
+                userPropertyDAO.delete(property);
+            }
+        } else {
+            property.setValue(tickerTaskAssignee);
+            userPropertyDAO.saveProp(property);
+        }
+    } catch (Exception e) {
+        // Return to the success page even though the pereference is not changed from default
+        // Avoid the error displays
+        request.setAttribute("status", "success");
+        return actionmapping.findForward("complete");
+    }
+    
 
 	    request.setAttribute("status", "success");
